+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Internal: Basic REST calls now return data directly. (no more `data` key). For advanced calls, use `call`.
     * Internal: Transfer SDK download is now a 2-step procedure: First get the YAML file from github with URLS for the various platforms and versions, and then download the archive from the official IBM repository.
     **global**: option `format=multi` is replaced with option `multi_table=yes`
+    * `faspex5`: removed deprecated option `value` replaced with positional parameter.
   
 * 4.19.0
   * New Features: